requirements.txt
aiohttp==3.8.4
SwissArmyTransformer==0.3.7
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
mdtex2html==1.2.0
fastapi==0.95.1
uvicorn==0.22.0
accelerate==0.20.3
python-dotenv==1.0.0